+++ awint_dag/src/common/lineage.rs
@@ -5,7 +5,7 @@ use crate::PState;
 /// A trait for mimicking structs that allows access to the internal state
 pub trait Lineage {
     fn state_nzbw(&self) -> NonZeroUsize {
-        self.state().get_nzbw().unwrap()
+        self.state().get_nzbw()
     }
 
     /// Get a reference to the `State` of `self`