+++ awint_dag/src/mimick/result.rs
@@ -8,7 +8,7 @@ use std::{
 use awint_ext::{awi, awint_internals::Location};
 use StdResult::{Err as StdErr, Ok as StdOk};
 
-use crate::{common::register_assertion_bit, dag, mimick::*};
+use crate::{dag, epoch::register_assertion_bit_for_current_epoch, mimick::*};
 
 // the type itself must be public, but nothing else about it can
 #[derive(Debug, Clone, Copy)]
@@ -278,7 +278,7 @@ impl<T, E> Result<T, E> {
                     line: tmp.line(),
                     col: tmp.column(),
                 };
-                register_assertion_bit(z.is_ok, location);
+                register_assertion_bit_for_current_epoch(z.is_ok, location);
                 if let StdOk(t) = z.res {
                     t
                 } else {
@@ -303,7 +303,7 @@ impl<T, E> Result<T, E> {
                     line: tmp.line(),
                     col: tmp.column(),
                 };
-                register_assertion_bit(!z.is_ok, location);
+                register_assertion_bit_for_current_epoch(!z.is_ok, location);
                 if let StdErr(e) = z.res {
                     e
                 } else {
@@ -376,7 +376,7 @@ impl<T, E> std::ops::Try for Result<T, E> {
                         line: tmp.line(),
                         col: tmp.column(),
                     };
-                    register_assertion_bit(z.is_ok, location);
+                    register_assertion_bit_for_current_epoch(z.is_ok, location);
                     ControlFlow::Continue(t)
                 }
                 StdErr(e) => {
@@ -386,7 +386,7 @@ impl<T, E> std::ops::Try for Result<T, E> {
                         line: tmp.line(),
                         col: tmp.column(),
                     };
-                    register_assertion_bit(!z.is_ok, location);
+                    register_assertion_bit_for_current_epoch(!z.is_ok, location);
                     ControlFlow::Break(Err(e))
                 }
             },