+++ awint_internals/src/lib.rs
@@ -166,7 +166,8 @@ pub const fn total_digits(w: NonZeroUsize) -> NonZeroUsize {
     unsafe { NonZeroUsize::new_unchecked(digits(w).wrapping_add((extra(w) != 0) as usize)) }
 }
 
-/// Location for an item in the source code
+/// Location for an item in the source code. This is essentially a
+/// `std::panic::Location<'static>` with all public fields.
 #[derive(Debug, Clone, Copy)]
 pub struct Location {
     pub file: &'static str,